+++ tests/mock_responses.py
@@ -12,8 +12,41 @@ camera_response = {
         "name": "API MOCK Camera",
         "serial": "E38D-BAEB-67C8",
         "site": "API MOCK Site",
+        "last_online": 200,
+        "status": "Live"
         }
     ]
+}
+camera_object_response = {
+    "object_counts": [
+        {
+            "people_count": 1,
+            "detected_time": 1595287853,
+            "vehicle_count": 0
+        },
+        {
+            "people_count": 8,
+            "detected_time": 1595287872,
+            "vehicle_count": 4
+        }
+    ],
+    "page_cursor": "go_to_two"
+}
+
+camera_object_response_two = {
+    "object_counts": [
+        {
+            "people_count": 2,
+            "detected_time": 1595287853,
+            "vehicle_count": 0
+        },
+        {
+            "people_count": 8,
+            "detected_time": 1595287872,
+            "vehicle_count": 4
+        }
+    ],
+    "page_cursor": None
 }
 
 notification_response = {